Fix bf16 compatibility for Qwen/modern models
- visualization.py: .cpu().float().numpy() instead of .cpu().numpy() (bf16 unsupported by numpy)
- visualization.py: cast mask to bool before invert (bf16 mask fails on ~operator)
- experiment.py: same .cpu().float().numpy() pattern for all tensor→numpy conversions
- Verified: Qwen2.5-3B warmup + training runs on 5090 (ppl 2.30→2.25)

+++ utils/neural_plasticity/experiment.py
@@ -257,7 +257,7 @@ def plot_baseline_entropy(
     fig, (ax1, ax2) = plt.subplots(1, 2, figsize=figsize)
     
     # Plot entropy
-    entropy_data = entropy_values.detach().cpu().numpy()
+    entropy_data = entropy_values.detach().cpu().float().numpy()
     im1 = ax1.imshow(entropy_data, cmap="viridis", aspect="auto")
     fig.colorbar(im1, ax=ax1, label='Entropy')
     ax1.set_title('Head Entropy (Higher = Less Focused)')
@@ -268,7 +268,7 @@ def plot_baseline_entropy(
     im1.set_clim(0, max(0.1, entropy_data.max()))
     
     # Plot gradients
-    grad_data = grad_norm_values.detach().cpu().numpy()
+    grad_data = grad_norm_values.detach().cpu().float().numpy()
     im2 = ax2.imshow(grad_data, cmap="plasma", aspect="auto")
     fig.colorbar(im2, ax=ax2, label='Gradient Norm')
     ax2.set_title('Head Gradient Norms (Higher = More Learning)')
@@ -1454,7 +1454,7 @@ class NeuralPlasticityExperiment:
             return None
         
         # Extract current model attention
-        current_attentions = [attn.detach().cpu().numpy() for attn in outputs.attentions]
+        current_attentions = [attn.detach().cpu().float().numpy() for attn in outputs.attentions]
         
         # Store the current model state
         current_state = self.model.state_dict().copy()
@@ -1471,7 +1471,7 @@ class NeuralPlasticityExperiment:
                     baseline_outputs = self.model(input_ids=input_ids, output_attentions=True)
                     
                 if hasattr(baseline_outputs, 'attentions') and baseline_outputs.attentions:
-                    baseline_attentions = [attn.detach().cpu().numpy() for attn in baseline_outputs.attentions]
+                    baseline_attentions = [attn.detach().cpu().float().numpy() for attn in baseline_outputs.attentions]
                 
                 # Restore current model state
                 self.model.load_state_dict(current_state)
@@ -2078,12 +2078,12 @@ class NeuralPlasticityExperiment:
                 # Send entropy and gradient data if available
                 if "entropy_values" in pruning_results and isinstance(pruning_results["entropy_values"], torch.Tensor):
                     self.metrics_callback(cycle, {
-                        "entropy_values": pruning_results["entropy_values"].detach().cpu().numpy()
+                        "entropy_values": pruning_results["entropy_values"].detach().cpu().float().numpy()
                     })
                 
                 if "grad_norm_values" in pruning_results and isinstance(pruning_results["grad_norm_values"], torch.Tensor):
                     self.metrics_callback(cycle, {
-                        "grad_norm_values": pruning_results["grad_norm_values"].detach().cpu().numpy()
+                        "grad_norm_values": pruning_results["grad_norm_values"].detach().cpu().float().numpy()
                     })
             
             # Save the heatmap
